change results output from script to tbl-based

+++ data-clone-roll-out-script.sql
@@ -1,11 +1,29 @@
 #------------------------------------------------------------------------------------------------------------------------------
+# TABLES
+#------------------------------------------------------------------------------------------------------------------------------
+
+#Create table to house the names of tables to be INCLUDED (empty if all tables required)
+Drop Table if exists data_clone_inclusions;
+Create Table data_clone_inclusions 
+	( 
+	tbl_name varchar(100) NOT NULL
+	) Engine = InnoDB;
+	
+#Create table to house the names of tables to be EXCLUDED (empty if no specific exclusions required)
+Drop Table if exists data_clone_exclusions;
+Create Table data_clone_exclusions 
+	( 
+	tbl_name varchar(100) NOT NULL
+	) Engine = InnoDB;
+
+#------------------------------------------------------------------------------------------------------------------------------
 # PROCEDURES
 #------------------------------------------------------------------------------------------------------------------------------
 
 DELIMITER $$
 
 DROP PROCEDURE IF EXISTS clone_data$$
-CREATE PROCEDURE clone_data (IN schemaname varchar(50), IN rowlimit Int)
+CREATE PROCEDURE clone_data (IN schemaname varchar(50), IN rowlimit Int, IN includefieldname Boolean, IN tempcleanup Boolean)
 Begin
 
 Declare sql_statement varchar(10000);
@@ -17,6 +35,29 @@ Set @Tab = '\t';
 Set @format_script = False;
 Set @row_limit = rowlimit;
 
+#Create table to house the names of tables to be INCLUDED (empty if all tables required)
+Create Table if not exists data_clone_inclusions 
+	( 
+	tbl_name varchar(100) NOT NULL
+	) Engine = InnoDB;
+	
+#Create table to house the names of tables to be EXCLUDED (empty if no specific exclusions required)
+Create Table if not exists data_clone_exclusions 
+	( 
+	tbl_name varchar(100) NOT NULL
+	) Engine = InnoDB;
+	
+
+#Create table to house insert statements
+Drop Table if exists data_clone_results;
+Create Table data_clone_results 
+	( 
+	tbl_name varchar(100) NOT NULL,
+    row_number int NOT NULL,
+    ins_header varchar(10000) NOT NULL,
+    ins_statement varchar(10000) NOT NULL
+	) Engine = InnoDB;
+
 
 #Create temp table to house tables to script out
 Drop Table if exists temp_clone_table_list;
@@ -26,6 +67,7 @@ Create Table temp_clone_table_list
     t_name varchar(100) NOT NULL,
     r_count int NOT NULL,
     clone_used boolean NOT NULL,
+    clone_status varchar(20) NOT NULL,
 	Primary Key (id)
 	) Engine = InnoDB;
 
@@ -36,6 +78,7 @@ Create Table temp_clone_field_list
 	( 
 	id int AUTO_INCREMENT,
     f_name varchar(100) NOT NULL,
+	f_type varchar(50) NOT NULL,
 	Primary Key (id)
 	) Engine = InnoDB;
 
@@ -45,11 +88,13 @@ Insert Into temp_clone_table_list
     t_name
     ,r_count
     ,clone_used
+    ,clone_status
     )
 	Select 
 		TABLE_NAME
 		,TABLE_ROWS
         ,FALSE
+        ,'WAITING'
 	From 
 		INFORMATION_SCHEMA.TABLES
 	Where
@@ -57,7 +102,7 @@ Insert Into temp_clone_table_list
 	And
 		TABLE_TYPE = 'BASE TABLE'
 	And
-		TABLE_NAME Not In ('temp_clone_table','temp_clone_table_list', 'temp_clone_field_list')
+		TABLE_NAME Not In ('data_clone_results','data_clone_inclusions','data_clone_exclusions','temp_clone_table','temp_clone_table_list', 'temp_clone_field_list')
 	; 
 
 
@@ -70,7 +115,15 @@ From
 	temp_clone_table_list
 ;
 
-Set @script_tables = '';
+#Capture whether the inclusions table is being used
+Select
+	Count(*)
+Into
+	@inclusions
+From
+	data_clone_inclusions
+;
+
 Set @script_insert = '';
 Set @t = 0;
 Set @all_data_rows = '';
@@ -83,13 +136,29 @@ TableLoop: LOOP
     
          LEAVE TableLoop;
          
-    ELSE 
+    ELSE
 		
         Set @script_current_table = '';
         Set @auto_increment_field = '';
         
         #Set current table name
         Select t_name Into @current_table From temp_clone_table_list Where id = @t;
+        
+        If Exists (Select 1 From data_clone_exclusions Where tbl_name = @current_table) Then
+		
+			Update temp_clone_table_list Set clone_status = 'EXCLUDED' Where id = @t;
+            ITERATE TableLoop;
+        
+        ElseIf (Select r_count From temp_clone_table_list Where t_name = @current_table) = 0 Then
+        
+            Update temp_clone_table_list Set clone_status = 'IGNORED' Where id = @t;
+            ITERATE TableLoop;
+		
+		Else
+        
+			Update temp_clone_table_list Set clone_status = 'STARTED' Where id = @t;
+        
+        End If;
         
         Select COLUMN_NAME Into @auto_increment_field From INFORMATION_SCHEMA.COLUMNS Where TABLE_SCHEMA = schemaname And TABLE_NAME = @current_table And EXTRA = 'auto_increment';
         
@@ -121,11 +190,9 @@ TableLoop: LOOP
             #Set @current_table = 'temp_clone_table';            
             Set @auto_increment_field = 'temp_clone_id';
 			Update temp_clone_table_list Set clone_used = True Where id = @t;
-		
+	
         End If;
-        
-        #Select @current_table;
-        
+
         Select clone_used Into @use_clone_table From temp_clone_table_list Where id = @t;
         
         #Prepare field variables
@@ -135,9 +202,11 @@ TableLoop: LOOP
 		Insert Into temp_clone_field_list
 			(
 			f_name
+			,f_type
 			)
 		Select 
 			COLUMN_NAME
+			,DATA_TYPE
 		From 
 			INFORMATION_SCHEMA.COLUMNS
 		Where
@@ -158,7 +227,7 @@ TableLoop: LOOP
 		From
 			temp_clone_field_list
 		;
-         
+          
          #Set field variables for field loop
          Set @f = 0;
          Set @field_list = '';
@@ -173,7 +242,7 @@ TableLoop: LOOP
 				LEAVE FieldLoop;
 			
             Else
-				
+					
                 Select f_name Into @f_name From temp_clone_field_list Where id = @f;
                 
 				If @format_script = True Then 
@@ -183,24 +252,24 @@ TableLoop: LOOP
 					Set @field_list = Concat(@field_list,',`',@f_name,'`');
 				
 				End If;
-                
+           
                 ITERATE FieldLoop;
 			
             End If;
-         
+       
          END LOOP FieldLoop;
-         
+
 		#Adjust field variables to remove spurious characters (like commas)
         Set @field_list = Substring(@field_list,2,Length(@field_list));
 		
         #Prepare table for row-level scripting
-        
+      
         #Set the current table (use clone table if required)
 		Select Case @use_clone_table When True Then 'temp_clone_table' Else t_name End Into @t_name From temp_clone_table_list Where id = @t; 
 		
         If @format_script = True Then 
 			
-			Set @script_insert = Concat(@script_insert,'Insert Into `',@current_table,'`',@NewLine,@Tab,'(',@NewLine,@Tab,@field_list,')');
+			Set @script_insert = Concat('Insert Into `',@current_table,'`',@NewLine,@Tab,'(',@NewLine,@Tab,@field_list,')');
             
 		Else
         
@@ -214,16 +283,16 @@ TableLoop: LOOP
 		Execute SQLCommand;
         
         Set sql_statement = Concat('Select Max(',@auto_increment_field,') Into @max_increment From ',@t_name,';');
-		Select sql_statement Into @sql_statement;
+        Select sql_statement Into @sql_statement;
 		Prepare SQLCommand From @sql_statement;
 		Execute SQLCommand;
         
         Set @row_no = 1;
         Set @record_exists = True;
+       
+		DataLoop: Loop 
         
-		DataLoop: Loop
-            
-			Set sql_statement = Concat('Select Case When Count(*) < 1 Then False Else True End Into @record_exists From ',@t_name,' Where ',@auto_increment_field,' = ',@d,';');
+			Set sql_statement = Concat('Select Case When Count(*) < 1 Then False Else True End Into @record_exists From ',@t_name,' Where ',@auto_increment_field,' = ',@d,';');          
 			Select sql_statement Into @sql_statement;
 			Prepare SQLCommand From @sql_statement;
 			Execute SQLCommand;
@@ -237,7 +306,8 @@ TableLoop: LOOP
                 #Reset field variables for data field loop
 				Set @f = 0;
 				Set @f_name = '';
-                Set @current_data_row = Concat(@script_insert,@NewLine,'(Select ');
+                #Set @current_data_row = Concat(@script_insert,@NewLine,'(Select ');
+				Set @current_data_row = '(Select ';
                 
 				DataFieldLoop: Loop
 				
@@ -252,25 +322,49 @@ TableLoop: LOOP
 						Select f_name Into @f_name From temp_clone_field_list Where id = @f;
 
 					End If;
-                    
+                   
 					Set sql_statement = Concat('Select IfNull(Convert(',@f_name,',char),''CLONE TO NULL'') Into @field_value From ',@t_name,' Where ',@auto_increment_field,' = ',@d,';');
 
 					Select sql_statement Into @sql_statement;
 					Prepare SQLCommand From @sql_statement;
 					Execute SQLCommand;
                     
-					Set @current_data_row = Concat(@current_data_row,'''',@field_value,''' As ',@f_name,',');
+					If includefieldname = True Then
+						
+						Set @current_data_row = Concat(@current_data_row,'''',Replace(@field_value,'''',''''''),''' As ',@f_name,',');
+					
+					Else
+					
+						Set @current_data_row = Concat(@current_data_row,'''',Replace(@field_value,'''',''''''),''',');
+					
+					End If;
                     
                     Iterate DataFieldLoop;
 				
 				END LOOP DataFieldLoop;
                 
                 #Tidy up the record
+                Select Replace(@current_data_row,'''CLONE TO NULL''','NULL') Into @current_data_row;
+                Set @current_data_row = Left(@current_data_row,Length(@current_data_row)-1);
+				Set @current_data_row = Concat(@current_data_row,');');	
+                
+				#Insert results into table
+				Insert Into data_clone_results
+					(
+					tbl_name
+                    ,row_number
+					,ins_header
+					,ins_statement
+					)
+				Select
+					@current_table
+                    ,@row_no
+					,@script_insert
+					,@current_data_row
+				;
+                
                 Set @all_data_rows = Concat(@all_data_rows,@NewLine,@current_data_row);
-                Select Replace(@all_data_rows,'''CLONE TO NULL''','NULL') Into @all_data_rows;
-                Set @all_data_rows = Left(@all_data_rows,Length(@all_data_rows)-1);
-				Set @all_data_rows = Concat(@all_data_rows,');',@NewLine);
-				
+                
                 #Increment the physical row count of the table
                 Set @row_no = @row_no + 1;
 			End If;
@@ -280,29 +374,27 @@ TableLoop: LOOP
 
 		END LOOP DataLoop;
         
-        Set @script_tables = Concat(@script_current_table,@NewLine,@all_data_rows);
-        
         #Clean up
         Drop Table If Exists temp_clone_table;
         Set @script_insert = '';
+        Update temp_clone_table_list Set clone_status = 'COMPLETED' Where id = @t;
 		ITERATE TableLoop;
     END IF;
 END LOOP TableLoop;
 
 #Clean up
-Drop Table If Exists temp_clone_field_list;
-Drop Table If Exists temp_clone_table_list;
+If tempcleanup = True Then
+
+	Drop Table If Exists temp_clone_field_list;
+	Drop Table If Exists temp_clone_table_list;
+
+End If;
 
 Select 
-	#@script_insert
-	#,@all_data_rows
-    #,@d
-    @script_tables
-    #,@script_current_table
-    #,@field_list
-    #,@field_count
-    #,@t_name
-    #,@f_name
+	ins_header
+    ,ins_statement
+From
+	data_clone_results
 ;
 
 End$$
